Перенести код в целевой компонент SignInForm
